Add getUserDefaultKeys() method to use the default IdentityFiles
* ssh defaults to the files "id_rsa", and "id_dsa", when the options
IdentityFile and IdentitiesOnly are not set.
* if there is an IdentityFile option set, it will be picked up by the current
readUserSSHConfig() and this method will never be used.
* there is no method yet that checks for the option IdentitiesOnly which can
lead to incorrect usage of this default files. However, this is likely to be
a non-issue since if IdentitiesOnly is set, an IdentityFile is likely to be
set for this host, in which case it will be caught by readUserSSHConfig().