SSHSessionCreator: fix javadoc
Actually, it seems that "ssh-keyscan" does not hash the lines ever.
But it does output the host key in all supported schemes, not just the
preferred one -- whereas if you SSH to the host, it will store the key
only in the preferred scheme (e.g., ecdsa-sha2-nistp256), even though
JSch would like it in a different one (in particular: ssh-rsa).